Refactor indexing and document handling: introduce make_document_batch utility for efficient document batching, update index_docs to utilize batching, and clean up unused code in homer_inmem.py.
